+++ tests/core_impls_and_misc/iter.rs
@@ -16,3 +16,31 @@ fn clone() {
     assert_eq!(it2.len(), 1);
     assert_eq!(it2.next(), Some(true));
 }
+
+#[test]
+fn nth() {
+    let mut it = exhaust::Iter::<u8>::default();
+    assert_eq!(it.nth(100), Some(100));
+    assert_eq!(it.next(), Some(101));
+    assert_eq!(it.nth(1000), None);
+}
+
+#[test]
+fn nth_back() {
+    let mut it = exhaust::Iter::<u8>::default();
+    assert_eq!(it.nth_back(10), Some(245));
+    assert_eq!(it.next_back(), Some(244));
+    assert_eq!(it.nth_back(1000), None);
+}
+
+#[test]
+fn last_simple() {
+    assert_eq!(exhaust::Iter::<u8>::default().last(), Some(255));
+}
+
+#[test]
+fn last_mutated() {
+    let mut it = exhaust::Iter::<u8>::default();
+    it.next_back();
+    assert_eq!(it.last(), Some(254));
+}